+++ src/ft_tcp_socket.h
@@ -1,15 +1,16 @@
 #ifndef FT_TCP_SOCKET_H
 #define FT_TCP_SOCKET_H
 
+#include <stdint.h>
 #include <sys/types.h>
 
 struct ft_tcpSocket {
     int fd;
     char *host;
-    int port;
+    uint16_t port;
 };
 
-struct ft_tcpSocket *ft_tcpSocketCreate(char *host, unsigned short port);
+struct ft_tcpSocket *ft_tcpSocketCreate(char *host, uint16_t port);
 int ft_tcpSocketConnect(struct ft_tcpSocket *socket);
 ssize_t ft_tcpSocketReceive(struct ft_tcpSocket *socket, void *buffer, size_t nBytes);
 ssize_t ft_tcpSocketSend(struct ft_tcpSocket *socket, const void *buffer, size_t nBytes);
